pyproject.toml
[project]
name = "in5"
version = "0.1.0"
description = ""
authors = [
    {name = "K Hemanth",email = "konankihemanth03@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "google-genai (>=1.28.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (<4.0)",
    "jwt (>=1.4.0,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
